+++ sl_run_checks.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   run_checks.c                                       :+:      :+:    :+:   */
+/*   sl_run_checks.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mmakagon <mmakagon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/16 15:35:24 by mmakagon          #+#    #+#             */
-/*   Updated: 2023/10/24 15:45:27 by mmakagon         ###   ########.fr       */
+/*   Updated: 2023/10/30 14:09:07 by mmakagon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,7 +47,7 @@ int	check_map_size(t_map *map, char *filename)
 		if (line == NULL)
 			break ;
 		map -> blocks_y += 1;
-		if (map->blocks_x != ft_strlen(line))
+		if (map->blocks_x != (int)ft_strlen(line))
 		{
 			close (fd);
 			return (free(line), MLX_ERROR);
@@ -64,23 +64,22 @@ int	check_map_contents(t_map *map, char *filename)
 	int		check;
 	char	*whole_map;
 
-
-	whole_map = (char *)ft_calloc((map->blocks_x + 1) * map->blocks_y + 1, sizeof(char));
+	whole_map = (char *)ft_calloc
+		((map->blocks_x + 1) * map->blocks_y + 1, sizeof(char));
 	fd = open(filename, O_RDONLY);
 	read(fd, whole_map, (map->blocks_x + 1) * map->blocks_y);
 	if (!ft_strchr(whole_map, '1') || !ft_strchr(whole_map, 'P')
-			|| !ft_strchr(whole_map, 'C') || !ft_strchr(whole_map, 'E'))
+		|| !ft_strchr(whole_map, 'C') || !ft_strchr(whole_map, 'E'))
 		check = MLX_ERROR;
 	else
 		check = 0;
 	free (whole_map);
 	return (check);
-
 }
 
 int	run_checks(int argc, t_map *map, char *filename)
 {
-    check_input (argc);
+	check_input (argc);
 	if ((check_map_size (map, filename)) == MLX_ERROR)
 		return (ft_printf ("Wrong map, check lines' lenght"), MLX_ERROR);
 	if (map->blocks_x < 5 || map->blocks_y < 3)
